refactor(web/sidebar): implement new sidebar navigation components
- Created AppSidebar component to serve as the main sidebar navigation.
- Added NavMain, NavProjects, NavUser components for organizing main navigation, project links, and user information respectively.
- Updated import paths for AppSidebar to reflect new structure.
- Introduced nav-chats component as a placeholder for future chat-related navigation.
