+++ apps/web/components/main-chat/main-chat.tsx
@@ -2,7 +2,7 @@
 
 import { CopilotChat } from "@copilotkit/react-ui";
 import { useCopilotAction } from "@copilotkit/react-core";
-import { AppSidebar } from "@/components/app-sidebar";
+import { AppSidebar } from "@/components/sidebar-nav/app-sidebar";
 import {
   Breadcrumb,
   BreadcrumbItem,
