+++ apps/web/components/sidebar-nav/app-sidebar.tsx
@@ -3,9 +3,9 @@
 import * as React from "react";
 import { Frame, Settings2 } from "lucide-react";
 
-import { NavMain } from "@/components/nav-main";
-import { NavProjects } from "@/components/nav-projects";
-import { NavUser } from "@/components/nav-user";
+import { NavMain } from "@/components/sidebar-nav/nav-main";
+import { NavProjects } from "@/components/sidebar-nav/nav-projects";
+import { NavUser } from "@/components/sidebar-nav/nav-user";
 import { PicoLogo } from "@/components/pico-logo";
 import {
   Sidebar,
